moved config and secret funcs in one file
